* include/stdint.h: Use int rather than long for {u,}int32_t.

+++ winsup/cygwin/include/stdint.h
@@ -17,7 +17,7 @@ details. */
 #define __int8_t_defined
 typedef signed char int8_t;
 typedef short int16_t;
-typedef long int32_t;
+typedef int int32_t;
 typedef long long int64_t;
 #endif
 
@@ -25,7 +25,7 @@ typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 #ifndef __uint32_t_defined
 #define __uint32_t_defined
-typedef unsigned long uint32_t;
+typedef unsigned int uint32_t;
 #endif
 typedef unsigned long long uint64_t;
